Changes to backend (mostly)
Front-end has been changed but not much (addition of initialise function and using _universalFood.js as the source of the Vue instance.

The key changes are to the backend. sql_retrieve.php has been greatly simplified, which is possible due to renaming of some variables in _universalFood.js. Now the food type pages can display options for each question dynamically based on food type (stored in session).