Label for images can now be clicked on
Previously, due to usage of <p>, images cannot be clicked on. Thus, I added a btn class to the <p> element in categories.html. However, there seems to be some sort of conflict between the styling for <p> and the Bootstrap styling for the btn class, so I had to add the exact same style for the p elements to the btn class.